% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AminoAcidFilter}
\alias{AminoAcidFilter}
\title{Amino acid parameter matrix.}
\format{
A matrix with 22 rows and 6 columns:
}
\usage{
AminoAcidFilter
}
\description{
A matrix containing properties of amino acids.
}
\keyword{datasets}
