% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLOSUMCappedDiff}
\alias{BLOSUMCappedDiff}
\title{Modified BLOSUM62 matrix.}
\format{
A matrix with 22 rows and 22 columns:
}
\usage{
BLOSUMCappedDiff
}
\description{
A matrix containing thresholded BLOSUM values to calculate distance between k-mers.
}
\keyword{datasets}
