% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPANTCR_Main.R
\docType{class}
\name{CDR3Breakdown-class}
\alias{CDR3Breakdown-class}
\alias{CDR3Breakdown}
\title{Set CDR3Breakdown class}
\description{
Set CDR3Breakdown class
}
\section{Slots}{

\describe{
\item{\code{CloneData}}{data.table.}

\item{\code{Elements}}{character.}

\item{\code{Metric}}{numeric.}

\item{\code{Output}}{data.table.}

\item{\code{Metadata}}{data.table.}
}}

