% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDR3SetDistance.R
\name{ComparisonFunction}
\alias{ComparisonFunction}
\title{Comparison function between two sets of TCRs generated by SPANTCR}
\usage{
ComparisonFunction(
  baseset,
  compset,
  limit = 100,
  ticks = FineBinTicks,
  functionmatrix = BLOSUMCappedDiff,
  fulloutput = F
)
}
\arguments{
\item{baseset}{CDR3Breakdown object 1}

\item{compset}{CDR3Breakdown object 2}

\item{limit}{Cumulative distance at which to stop comparing CDR3s}

\item{ticks}{Number of ticks to use, default 100}

\item{functionmatrix}{Matrix of amino acid distance metrics, default modified BLOSUM}

\item{fulloutput}{T/F if all TCRs in compset should be saved}
}
\value{
List containing the cumulative distance between each TCR of base set and either the top closest k-mers in compset or all k-mers
}
\description{
Comparison function between two sets of TCRs generated by SPANTCR
}
