% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropyAnalysisFunctions.R
\name{EntropyScan}
\alias{EntropyScan}
\title{Checks entropy contributions for each major motif in each range window}
\usage{
EntropyScan(Data, Searches, Limit = 0.03, ticksused = FineBinTicks)
}
\arguments{
\item{Data}{CDR3Breakdown object}

\item{Searches}{List of vectors containing search ranges, default SearchBoxes for single CDR3 SearchBoxesPaired for paired}

\item{Limit}{Minimum frequency of k-mers to analyze in each bin, default 0.03}

\item{ticksused}{Vector of ticks to analyze, FineBinTicks for single CDR3, FineBinTicksPaired for paired CDR3s}
}
\value{
Summarized result of function SearchIterator called on each range in Searches
}
\description{
Checks entropy contributions for each major motif in each range window
}
