% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FineBinTicks}
\alias{FineBinTicks}
\title{Vector of tick centroids 1-100 to use with single CDR3 genes for distance comparisons.}
\format{
A vector of length 100:
}
\usage{
FineBinTicks
}
\description{
A vector from 0.005 to 0.995.
}
\keyword{datasets}
