% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FineBinTicksPaired}
\alias{FineBinTicksPaired}
\title{Vector of tick centroids 1-200 to use with paired CDR3 for distance comparisons.}
\format{
A vector of length 200:
}
\usage{
FineBinTicksPaired
}
\description{
A vector from 0.005 to 1.995.
}
\keyword{datasets}
