% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPANTCR_Main.R
\name{ScoreFunctionExponential5}
\alias{ScoreFunctionExponential5}
\title{TCR score functions}
\usage{
ScoreFunctionExponential5(x)
}
\arguments{
\item{x}{Vector of TCR scores}
}
\description{
TCR score functions
}
