% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPANTCR_Main.R
\name{ScoreFunctionSquared}
\alias{ScoreFunctionSquared}
\title{TCR input score function}
\usage{
ScoreFunctionSquared(x)
}
\arguments{
\item{x}{Vector of TCR scores}
}
\description{
TCR input score function
}
