% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropyAnalysisFunctions.R
\name{SearchIterator}
\alias{SearchIterator}
\title{Checks entropy contributions for all TCRs expressing a specific motif (searchterms) in a specific range (targetrange)}
\usage{
SearchIterator(base, searchterms, targetrange, ticks = FineBinTicks)
}
\arguments{
\item{base}{CDR3Breakdown object}

\item{searchterms}{Vector of k-mers to analyze matching the length analyzed in base}

\item{targetrange}{Two element vector describing the lower and upper bound of CDR3 region analyzed (from 0-1)}

\item{ticks}{Vector of ticks to analyze, FineBinTicks for single CDR3, FineBinTicksPaired for paired CDR3s}
}
\value{
data.table object describing all k-mers contribution to entropy
}
\description{
Checks entropy contributions for all TCRs expressing a specific motif (searchterms) in a specific range (targetrange)
}
