% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPANTCR_Main.R
\name{SPANTCR}
\alias{SPANTCR}
\title{Analyze a TCR Set}
\usage{
SPANTCR(
  Data,
  Filename,
  FileType,
  AAOperation = "Hydrophobicity",
  TickNumber = 100,
  SlidingWindowTickSize = 2,
  SignificanceCutoff = 0.03,
  ScoreFunction = ScoreFunctionExponential5,
  WeightFunction = WeightFunctionLinear
)
}
\arguments{
\item{Data}{Datatable with columns CDR3/vgene/jgene/gene/score/id (if paired). Column gene has values "TCRA" or "TCRB". Column id has matched numbers for paired TCRA/B genes.}

\item{Filename}{Name of file or data}

\item{FileType}{"Paired" "TCRA" or "TCRB"}

\item{AAOperation}{Column to use from object AminoAcidFilter}

\item{TickNumber}{Number of bins to use, default 100}

\item{SlidingWindowTickSize}{k-mer length, 1/2/3 allowed currently, default 2}

\item{SignificanceCutoff}{Minimum proportion of k-mers present in each bin to be displayed, default 0.03}

\item{ScoreFunction}{How to prioritize internal scoring of TCRs, default 5^score}

\item{WeightFunction}{How to weight the influence of each k-mer, default linear growth/decay}
}
\value{
CDR3Breakdown object, data table of CDR3 composition in bins
}
\description{
Take a set of TCR CDR3s and process into a CDR3Breakdown object
}
\examples{
SPANTCR(VDJDBCMVPairedInput, "VDJDBCMV", "Paired", "Hydrophobicity", 100, 2,
0.03, ScoreFunctionExponential5, WeightFunctionLinear)
SPANTCR(VDJDBCMVPairedInput[gene=="TRA"], "VDJDBCMVAlpha","TRA","Mass", 100,
2, 0.05, ScoreFunctionExponential5, WeightFunctionLinear)
}
