% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VDJDBCMVPairedInput}
\alias{VDJDBCMVPairedInput}
\title{Sample data table of paired VDJDB CMV pp65 epitope data}
\format{
A data.table with 444 rows and 6 columns:
}
\usage{
VDJDBCMVPairedInput
}
\description{
A data.table properly formatted for SPAN-TCR
}
\keyword{datasets}
